OutOfMemory crash fix * open any chat * choose a video (i tested with videos at least) * write a text * add a spoiler * hide the app * return the app.. wtf crash.
It happends on pixel7 a13 and not on mi mix2 under los a11.

original flow is fucked up. Text layout return 1048576 size while in background so it fucks that app (cause it want 200+Mb momentally for a fucking arraylist with long (8b each!).

I added a fix (thats more than 16k screen), which totally remove that crash. But TG-code-monkeys need to fix the flow as is to solve the problem.